pyproject.toml
[tool.poetry]
name = "discord-cppref"
version = "0.1.0"
description = ""
authors = ["Ethan Rickert <ethanrickert0429@gmail.com>"]
readme = "README.md"
# packages = [{include = "discord_cppref"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.1"
"discord.py" = "2.0.0"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
